lib/cloudfunctions/price_updater/requirements.txt
functions-framework==3.5.0
psycopg2-binary==2.9.9
yfinance==0.2.33
pandas==2.0.3
numpy==1.24.3
google-cloud-logging==3.8.0